Move withDbHelpers to the separate file
