+++ app/support/DbAdapter/utils.js
@@ -71,47 +71,6 @@ export function sqlIntarrayIn(field, list) {
   return pgFormat(`(${list.inclusive ? '' : 'not '}${field} && %L)`, `{${list.items.join(',')}}`);
 }
 
-const dbHelpers = {
-  async getAll(sql, args = {}) {
-    const { rows } = await this.raw(sql, args);
-    return rows;
-  },
-
-  async getRow(sql, args = {}) {
-    const rows = await this.getAll(sql, args);
-    return rows[0];
-  },
-
-  async getOne(sql, args = {}, column = 0) {
-    const cols = await this.getCol(sql, args, column);
-    return cols[0];
-  },
-
-  async getCol(sql, args = {}, column = 0) {
-    const { rows, fields } = await this.raw(sql, args);
-
-    if (typeof column === 'number') {
-      column = fields[column].name;
-    }
-
-    return rows.map((r) => r[column]);
-  },
-
-  transaction(action) {
-    // eslint-disable-next-line prefer-reflect
-    return Object.getPrototypeOf(this).transaction((trx) => action(withDbHelpers(trx)));
-  },
-};
-
-export function withDbHelpers(db) {
-  // db is a function with additional properties, so extending is tricky
-  const wrapper = Object.assign(function (...args) {
-    return db.apply(this, args); // eslint-disable-line prefer-reflect
-  }, dbHelpers);
-  Object.setPrototypeOf(wrapper, db); // eslint-disable-line prefer-reflect
-  return wrapper;
-}
-
 function joinThem(array, joinBy, defaultValue, shortcutValue, skipValue) {
   if (array.some((x) => x === shortcutValue)) {
     return shortcutValue;
